+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Method to get the available CAN devices.
 - GPIO module.
 - Optional password for the FOE bootloader.
+- Method to check if drive is connected via a COM-KIT.
 
 ### Fix
 - Bug retrieving interface adapter name in Linux.